Made Admin Adding route and prevented users from making themselves admins
