+++ server/app/routes/index.js
@@ -4,7 +4,7 @@ module.exports = router;
 
 router.use('/members', require('./members'));
 router.use('/cart', require('./cart'));
-router.use('/users' , require('./users'));
+router.use('/user' , require('./user'));
 router.use('/products', require('./products'));
 router.use('/orders',require('./orders'));
 router.use('/admin',require('./admin'));
